Added tests for the line intersection
I am highly impressed with the value of implementing standard algorithms
as opposed to winging it. I tested a bunch of corner cases that I
probably wouldn't have even thought about if I'd have tried to
come up with a line intersection algorithm from scratch. And the
code handled all but one correctly - an extremely minor bug where two
colinear points that touch on one end would incorrectly be considered
non-colliding. And that bug was a tweak of mine to the stated algorithm
for implementation expediency