Astroid children now vary less in angle
In other words, when astroids explode, the children now tend to share
the same angle. They use a normal offset of the parent's angle, instead
of having uniformly distributed angles. This isn't part of the original
astroids game, to my knowledge. It just seemed like it made more sense
to have astroids continue to travel in roughly the same direction after
they exploded
That said, the effect doesn't work quite as intended, so it could use a
bit more tweaking