Add dot and cross product tests + minor refactor
Adds test for Point::dot and Point::cross product
The test case uses prime numbers instead of 0s and 1s so that it's more
robust against minor types. This case of test case brittleness was
discovered through manual fuzzing
Includes a refactor of the Point::new test to minimize duplication of
code fixtures. The exact test case was changed as well to enable this